V50: BattleCard audit fixes + new features + cleanup
BattleCard system: per-instance tween IDs (P1), frozen animation fix (P0),
correct achievement IDs for Earn cards (P1), deduplicate MatchmakingService
BattleCardId read (P2), add LoadFromFirebase for multi-device sync (P2),
fix SyncToFirebase missing userId, warn on 0 assets loaded (P3).
New: Win effects, LevelUp panel, Payments system, Cosmetics (BattleCards +
Titles + Frames), Navigation system, 12 new themes, iOS bridges (ATT +
StoreKit), Firebase Functions, database rules. Remove obsolete docs/assets.

+++ Assets/_Project/Scripts/Runtime/Features/CashBattle/Hub/CashBattle1v1Manager.cs
@@ -63,7 +63,7 @@ namespace DigitPark.UI.CashBattle
         // Constants - Triumph SDK limits
         private const decimal MIN_ENTRY_FEE = 1m;
         private const decimal MAX_ENTRY_FEE = 250m;
-        private const float WINNER_PERCENTAGE = 0.70f; // 70% to winner (Triumph takes 20%, platform 10%)
+        private const decimal WINNER_PERCENTAGE = 0.70m; // 70% to winner (Triumph takes 20%, platform 10%)
 
         // State
         private GameType? selectedGame;
@@ -561,7 +561,7 @@ namespace DigitPark.UI.CashBattle
         {
             // Winner gets 70% of total pot (entry fee x 2)
             // Triumph takes 20%, platform takes 10%
-            return entryFee * 2 * (decimal)WINNER_PERCENTAGE;
+            return entryFee * 2 * WINNER_PERCENTAGE;
         }
 
         private void UpdateSelectedFeeDisplay()
